Created ManifestHelper and moved AndroidManifest metadata read methods from SugarConfig to ManifestHelper
Reason: SugarConfig seemed to have 2 different responsibilities (holding
the field cache and retrieving data from the AndroidManifest), this
separates those concerns into 2 different classes.

Changes

1. Created ManifestHelper
   ManifestHelper provides methods for accessing the AndroidManifest's
   metadata (i.e. the database version, database name, etc.)
2. Added documentation to ManifestHelper
3. Refactored references that used SugarConfig to use ManifestHelper

Note:
There was an integer Flag with value 128, I'm not sure what that means
so I added a TODO.
